+++ src/missions_lib/__init__.py
@@ -1,5 +1,7 @@
 # __init__.py
 from .greet_all import Mission2StateMachine
+from .object_search import Mission3StateMachine
+